+++ queries/paste_back.py
@@ -8,7 +8,6 @@ class PasteBack(InsertionQuery):
 	def handle_single(self,view_information,query_description,extra = {}):
 		state = extra["state"]
 		history  =  extra["history"]
-		code = view_information["code"]
 		index = len(history)
 		while history[index-1][0]=="selection"  and index>=1 and history[index-1][1]  == view_information["change_count"]:
 			index -=1
@@ -16,19 +15,14 @@ class PasteBack(InsertionQuery):
 			return []
 		selection = history[index][2]
 		selection = selection if isinstance(selection,list) else [selection]
-		f = query_description["format"]
-		if f==1:
-			i = query_description["paste_back_index"]
-		elif f==2:
-			i = query_description["color"]
-		else:
-			return []
-		result = state["result"]
-		alternatives = state["alternatives"]
-		location = alternatives[i-1] if i != 0 else result
-		location = location if isinstance(location,list) else [location]
+		i = query_description.get("color",0)
+		print(" i.e.'s",i)
+		result_text = state["result_text"]
+		alternatives_text = state["alternatives_text"]
+		output = alternatives_text[i-1] if i != 0 else result_text
 		surrounding = query_description.get("surrounding_punctuation",("",""))
-		return [(x,surrounding[0]+code[l[0]:l[1]]+surrounding[1])  for x,l in zip(selection, location)]
+		print("output",output)
+		return [(x,surrounding[0]+output+surrounding[1])  for x in selection]